virtio-vhost-user: move PCI-specific code into virtio-vhost-user-pci.c
The code for setting up the PCIMemoryRegions and the virtio PCI
capabilities for the additional device resources (doorbells,
notifications, shared memory) is PCI transport specific. Move this code
along with the read/write handlers for the MemoryRegions into
hw/virtio/virtio-vhost-user-pci.c. The setting up of the virtio
capabilities and the associated PCIMemoryRegions becomes part of the
realization of the TYPE_VIRTIO_VHOST_USER_PCI object.

Signed-off-by: Nikos Dragazis <ndragazis@arrikto.com>