virtio-vhost-user: fix bug during termination of master VM
The problem is that the master QEMU process hangs after guest shutdown.
This patch fixes this bug.

This patch incorporates the following changes:

1. parse GET_VRING_BASE messages
According to the vhost-user protocol specification, the master sends
GET_VRING_BASE messages to indicate the stop of the vring operations.
The virtio-vhost-user device should parse those messages and cleanup the
kickfds/callfds associated with the vq_idx of each message.

2. change the cleanup procedure for the kickfds/callfds
The problem with the event_notifier_cleanup() function is that it does
not reset the rfd in case the rfd and the wfd have the same value.
However, the event_notifier_get_fd() always returns the rfd value. So,
instead of using event_notifier_cleanup(), we have to manually close the
rfd eventfd and reset its value to -1.

3. set the kickfds into non-blocking mode, just like the callfds
Otherwise, the master QEMU process hangs during VM shutdown.

Signed-off-by: Nikos Dragazis <ndragazis@arrikto.com>